+++ android/src/main/java/com/amplitude/android/utilities/AndroidNetworkConnectivityChecker.kt
@@ -35,21 +35,33 @@ class AndroidNetworkConnectivityChecker(private val context: Context, private va
             return true
         }
 
-        val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE)
-        if (cm is ConnectivityManager) {
-            if (isMarshmallowAndAbove) {
-                val network = cm.activeNetwork ?: return false
-                val capabilities = cm.getNetworkCapabilities(network) ?: return false
+        try {
+            val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE)
+            if (cm is ConnectivityManager) {
+                if (isMarshmallowAndAbove) {
+                    val network = cm.activeNetwork ?: return false
+                    val capabilities = cm.getNetworkCapabilities(network) ?: return false
 
-                return capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) ||
-                    capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
+                    return capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) ||
+                        capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
+                } else {
+                    @SuppressLint("MissingPermission")
+                    val networkInfo = cm.activeNetworkInfo
+                    return networkInfo != null && networkInfo.isConnectedOrConnecting
+                }
             } else {
-                @SuppressLint("MissingPermission")
-                val networkInfo = cm.activeNetworkInfo
-                return networkInfo != null && networkInfo.isConnectedOrConnecting
+                logger.debug("Service is not an instance of ConnectivityManager. Offline mode is not supported")
+                return true
             }
-        } else {
-            logger.debug("Service is not an instance of ConnectivityManager. Offline mode is not supported")
+        } catch (throwable: Throwable) {
+            // We've seen issues where we see exceptions being thrown by connectivity manager
+            // which crashes an app. Its safe to ignore these exceptions since we try our best
+            // to mark a device as offline
+            // Github Issues:
+            // https://github.com/amplitude/Amplitude-Kotlin/issues/220
+            // https://github.com/amplitude/Amplitude-Kotlin/issues/197
+            logger.warn("Error checking network connectivity: ${throwable.message}")
+            logger.warn(throwable.stackTraceToString())
             return true
         }
     }